+++ app/src/main/java/com/rama/mako/activities/settings/SettingsCheckboxController.kt
@@ -6,6 +6,7 @@ import android.content.Intent
 import android.view.View
 import android.widget.Toast
 import com.rama.mako.R
+import com.rama.mako.activities.MainActivity
 import com.rama.mako.activities.SettingsActivity
 import com.rama.mako.managers.PrefsManager.PrefKeys
 import com.rama.mako.receivers.ScreenLockAdminReceiver
@@ -49,8 +50,16 @@ class SettingsCheckboxController(private val activity: SettingsActivity) {
         bindWdCheckbox(R.id.show_battery_temperature, PrefKeys.BATTERY_TEMPERATURE, false)
         bindWdCheckbox(R.id.show_battery_charge_status, PrefKeys.BATTERY_CHARGE_STATUS, false)
         bindWdCheckbox(R.id.show_system_bar, PrefKeys.SYSTEM_BAR_VISIBLE, false)
-        setupDoubleTapToSleepCheckbox()
+        bindWdCheckbox(
+            R.id.prevent_home_screen_rotation,
+            PrefKeys.SYSTEM_PREVENT_ROTATION,
+            false,
+            onChange = { isChecked ->
+                activity.applyRotationLock(isChecked)
+            }
+        )
         bindWdCheckbox(R.id.show_profile_indicator, PrefKeys.APPS_PROFILE_INDICATOR, true)
+        setupDoubleTapToSleepCheckbox()
 
         bindWdCheckbox(
             R.id.lock_settings,
@@ -63,6 +72,7 @@ class SettingsCheckboxController(private val activity: SettingsActivity) {
             PrefKeys.SECURITY_KEYPAD_RANDOMIZED,
             true,
         )
+
     }
 
     fun onActivityResult(requestCode: Int, resultCode: Int) {